Voice/pyproject.toml
[tool.poetry]
name = "philbot-voice"
version = "0.13.0"
description = ""
authors = ["philipp.lengauer <p.lengauer@gmail.com>"]
packages = [{include = "philbot-voice"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
websocket-client = "^1.4.2"
PyNaCl = "^1.5.0"
PyOgg = "0.6.14a1"
youtube_dl = "^2021.12.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
philbot-voice = "voice:main"
